fix: ajuste de erro NaN

+++ calculadora/src/main/Calculator.jsx
@@ -62,7 +62,7 @@ export default class Calculator extends Component {
 
         if (n !== '.') {
             const i = this.state.current
-            const newValue = parseInt(displayValue)
+            const newValue = parseFloat(displayValue)
             const values = [ ...this.state.values]
             values[i] = newValue
             this.setState({ values })
